// src/app/core/interfaces/IPojectParticipantDTO.ts
export interface IProjectParticipantDTO{
    anonimo: number;
    codFaculProgApoy: number;
    codProgramaApoyado: number;
    dedicacionHoras: number;
    dedicacionHorasPlan: number;
    dedicacionMeses: number;
    dedicacionMesesPlan: number;
    dedicacionNN: string;
    detalleContrato: string;
    esInvestigadorPrinci: string;
    esInvestigadorPrincipal: number;
    fechaActualiza: Date;
    fechaCrea: Date;
    fechaFin: Date;
    fechaInicio: Date;
    funcion: string;
    grupo: number;
    identificador: number;
    institucion: string;
    nombreCortoGrupo: string;
    nombreGrupo: string;
    nombrePersona: string;
    nombreProgramaExterno: string;
    observacion: string;
    participaBeneficios: string;
    personaNatural: string;
    porcentajeBeneficios: number;
    porcentajeProgAcad: number;
    proyecto: string;
    rolParticipante: string;
    rolParticipanteProyecto: number;
    selectorInstitucion: string;
    selectorPersona: string;
    tipoContrato: string;
    usuarioActualiza: string;
    usuarioCrea: string;
    vinculacionClase: string;
    vinculacionTipo: string;
    vinculacionUdea: string;
    correo: string;
   }